# heptacert/frontend/public/openapi-gpt.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HeptaCert API",
    "description": "End-to-end event management: create events, manage attendees, run QR check-in, issue certificates, trigger automation, and view analytics.",
    "version": "1.0.0"
  },
  "servers": [{ "url": "https://heptacert.com" }],
  "security": [{ "bearerAuth": [] }],
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/api/admin/events/{event_id}/registration-fields": {
      "get": {
        "operationId": "getRegistrationFields",
        "summary": "List all registration form fields for an event",
        "parameters": [{ "name": "event_id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Form fields array" } }
      },
      "post": {
        "operationId": "addRegistrationField",
        "summary": "Append a single field to the registration form",
        "parameters": [{ "name": "event_id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["label", "name"],
                "properties": {
                  "label": { "type": "string", "description": "Display label shown to registrant" },
                  "name": { "type": "string", "description": "Internal snake_case field name" },
                  "type": { "type": "string", "enum": ["text", "email", "tel", "textarea", "select", "checkbox", "number", "date"], "default": "text" },
                  "required": { "type": "boolean", "default": false },
                  "placeholder": { "type": "string" },
                  "options": { "type": "array", "items": { "type": "string" }, "description": "Only for select type" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Appended field" } }
      },
      "put": {
        "operationId": "replaceRegistrationFields",
        "summary": "Replace ALL registration form fields. Pass empty array to clear the form.",
        "parameters": [{ "name": "event_id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["fields"],
                "properties": {
                  "fields": { "type": "array", "items": { "type": "object" } }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Updated fields" } }
      }
    },
    "/api/admin/events/{event_id}/ticket-types": {
      "get": {
        "operationId": "listTicketTypes",
        "summary": "List ticket types (pricing tiers) for an event",
        "parameters": [{ "name": "event_id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Ticket type list" } }
      },
      "post": {
        "operationId": "createTicketType",
        "summary": "Create a ticket type (e.g. Early Bird 250 TRY, VIP 500 TRY)",
        "parameters": [{ "name": "event_id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string" },
                  "description": { "type": "string" },
                  "price": { "type": "number", "default": 0 },
                  "currency": { "type": "string", "default": "TRY" },
                  "capacity": { "type": "integer", "description": "Max tickets of this type, null = unlimited" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created ticket type" } }
      }
    },
    "/api/admin/events/{event_id}/ticket-types/{type_id}": {
      "patch": {
        "operationId": "updateTicketType",
        "summary": "Update a ticket type (price, capacity, active status)",
        "parameters": [
          { "name": "event_id", "in": "path", "required": true, "schema": { "type": "integer" } },
          { "name": "type_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "price": { "type": "number" },
                  "currency": { "type": "string" },
                  "capacity": { "type": "integer" },
                  "is_active": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": { "200": { "description": "Updated ticket type" } }
      },
      "delete": {
        "operationId": "deleteTicketType",
        "summary": "Delete a ticket type",
        "parameters": [
          { "name": "event_id", "in": "path", "required": true, "schema": { "type": "integer" } },
          { "name": "type_id", "in": "path", "required": true, "schema": { "type": "integer" } }
        ],
        "responses": { "200": { "description": "Deleted" } }
      }
    }
  }
}